requirements.txt
gradio==5.49.1
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0
langchain-mcp-adapters>=0.1.0
fastmcp>=0.1.0
python-dotenv>=1.0.0
chromadb>=0.4.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.0
httpx>=0.24.0
telethon>=1.34.0

